feat: SetupDataLoader 클래스로 분리
